# internal/openapi/spec/public.yaml
openapi: 3.0.3
info:
  title: metareel Public API
  version: "1.0.0"
  description: >
    Every /api/v1 endpoint needs an API key, sent as `X-API-Key: <key>` or
    `Authorization: Bearer <key>` (unless keys are turned off in the admin
    console). Keys are created and rotated in the admin console under
    Settings & API keys. Requests are rate limited per key; over the limit
    the API returns 429 with Retry-After. Missing or invalid keys get 401.
servers:
  - url: /
security:
  - ApiKeyHeader: []
  - BearerKey: []
paths:
...
  /api/v1/titles:
    get:
      operationId: listTitles
      summary: List titles (paged)
      parameters:
        - name: kind
          in: query
          required: false
          schema:
            type: string
            enum: [movie, tv_show]
        - name: q
          in: query
          required: false
          schema:
            type: string
            description: Name or slug substring
        - name: missing
          in: query
          required: false
          schema:
            type: string
            enum: [tmdb, imdb, rt]
            description: Only titles without this ID
        - name: match_status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, matched, unmatched, manual]
            description: Only titles in this matching state (the admin review queue uses unmatched)
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [id_desc, id_asc, name_asc, name_desc, updated_desc, updated_asc, last_ranked_desc, last_ranked_asc, rankings_desc, rankings_asc]
            description: Default id_desc (newest first)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
      responses:
        "200":
          description: Titles page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitlesPageResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/titles/{id}:
    get:
      operationId: getTitle
      summary: Get one scraped title
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Title
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/titles/{id}/rankings:
    get:
      operationId: getTitleRankingsById
      summary: Ranking history of one scraped title (mapped to TMDB or not)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: "#/components/parameters/HistoryCountryParam"
        - $ref: "#/components/parameters/FromParam"
        - $ref: "#/components/parameters/ToParam"
      responses:
        "200":
          description: Ranking history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleRankingsResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/titles/tmdb/{kind}/{id}/rankings:
    get:
      operationId: getTitleRankings
      summary: Ranking history of one title, looked up by TMDB ID
      description: >
        Every Top 10 chart appearance of the title, oldest first, plus a
        per-chart summary. All FlixPatrol titles mapped to the TMDB ID are
        included. Returns 404 if no scraped title is mapped to the TMDB ID.
      parameters:
        - $ref: "#/components/parameters/TmdbKindParam"
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: TMDB ID (numeric)
        - $ref: "#/components/parameters/HistoryCountryParam"
        - $ref: "#/components/parameters/FromParam"
        - $ref: "#/components/parameters/ToParam"
      responses:
        "200":
          description: Ranking history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleRankingsResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No title mapped to this TMDB ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/titles/tmdb/{kind}/{id}/ratings:
    get:
      operationId: getTitleRatings
      summary: IMDb and Rotten Tomatoes ratings of a title, looked up by TMDB ID
      description: >
        Served from the database when refreshed within RATINGS_TTL (12h by
        default), otherwise refreshed from JustWatch, MDBList (when
        MDBLIST_API_KEY is set) and Rotten Tomatoes' own search index. Rotten
        Tomatoes scores come from RT directly when it has a match (by stored
        slug, MDBList's RT link, or a title + year search), else from MDBList /
        JustWatch. Every rating each provider returned is listed in sources.
        If a refresh fails, the previous ratings are returned with stale=true.
      parameters:
        - $ref: "#/components/parameters/TmdbKindParam"
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: TMDB ID (numeric)
        - name: refresh
          in: query
          required: false
          schema:
            type: boolean
            description: Refresh from the providers even if the stored ratings are fresh
      responses:
        "200":
          description: Ratings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingsResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No provider knows this TMDB title
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Every provider failed and nothing was stored yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/titles/tmdb/rankings:
    get:
      operationId: getTitleRankingsBatch
      summary: Ranking history of several titles, looked up by TMDB ID
      description: >
        Batch version of getTitleRankings. Items come back in request order;
        IDs with no mapped title are listed in not_found.
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: string
            example: movie:425,tv:154385
            description: Comma-separated kind:tmdb_id pairs (kind is movie or tv), at most 50
        - $ref: "#/components/parameters/HistoryCountryParam"
        - $ref: "#/components/parameters/FromParam"
        - $ref: "#/components/parameters/ToParam"
      responses:
        "200":
          description: Ranking histories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleRankingsBatchResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/titles/new/{country}/{service}:
    get:
      operationId: getNewTitles
      summary: Titles added to a streaming service, per day (live from JustWatch)
      description: >
        One JustWatch query per day from `from` to `to` (default: the last 7
        days, at most 31). Newest day first. Seasons are listed under their
        show. Nothing is stored.
      parameters:
        - name: country
          in: path
          required: true
          schema:
            type: string
            description: ISO country code (e.g. US)
        - name: service
          in: path
          required: true
          schema:
            type: string
            description: >
              FlixPatrol slug (netflix, hbo-max, disney, amazon-prime,
              paramount-plus, peacock, apple-tv; resolved to the country's
              JustWatch code) or any JustWatch package short / technical name
              from /api/v1/services/{country} (e.g. nfx, ppa)
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [movie, tv]
        - name: language
          in: query
          required: false
          schema:
            type: string
            default: en
            description: 2-letter language for titles, descriptions and genres
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
            description: Default today (UTC)
      responses:
        "200":
          description: New titles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleasesResponse"
        "400":
          description: Validation error (country, service, dates, type)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: JustWatch request failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/titles/upcoming/{country}/{service}:
    get:
      operationId: getUpcomingTitles
      summary: Titles announced for a streaming service (live from JustWatch)
      description: >
        Soonest first. date is the title's release date on this service
        (digital release preferred), not its theatrical or original premiere.
        Titles without a date on the service come last, and only when no
        from / to range is given. Nothing is stored.
      parameters:
        - name: country
          in: path
          required: true
          schema:
            type: string
            description: ISO country code (e.g. US)
        - name: service
          in: path
          required: true
          schema:
            type: string
            description: >
              FlixPatrol slug (netflix, hbo-max, disney, amazon-prime,
              paramount-plus, peacock, apple-tv; resolved to the country's
              JustWatch code) or any JustWatch package short / technical name
              from /api/v1/services/{country} (e.g. nfx, ppa)
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [movie, tv]
        - name: language
          in: query
          required: false
          schema:
            type: string
            default: en
            description: 2-letter language for titles, descriptions and genres
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
            description: Earliest release date on the service
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
            description: Latest release date on the service
      responses:
        "200":
          description: Upcoming titles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleasesResponse"
        "400":
          description: Validation error (country, service, dates, type)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: JustWatch request failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/services/{country}:
    get:
      operationId: listServices
      summary: Streaming services JustWatch knows in a country
      parameters:
        - name: country
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServicesResponse"
        "400":
          description: Validation error (country, service, dates, type)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: JustWatch request failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/top10/movies/{country}:
    get:
      operationId: getTop10MoviesAllProviders
      summary: Get Top 10 movies across all providers
      parameters:
        - $ref: "#/components/parameters/CountryParam"
        - $ref: "#/components/parameters/DateParam"
      responses:
        "200":
          description: Top 10 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Top10Response"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No data found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/top10/movies/{country}/{provider}:
    get:
      operationId: getTop10MoviesByProvider
      summary: Get Top 10 movies for a provider
      parameters:
        - $ref: "#/components/parameters/CountryParam"
        - $ref: "#/components/parameters/ProviderParam"
        - $ref: "#/components/parameters/DateParam"
      responses:
        "200":
          description: Top 10 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Top10Response"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No data found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/top10/tv-shows/{country}:
    get:
      operationId: getTop10TVShowsAllProviders
      summary: Get Top 10 TV shows across all providers
      parameters:
        - $ref: "#/components/parameters/CountryParam"
        - $ref: "#/components/parameters/DateParam"
      responses:
        "200":
          description: Top 10 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Top10Response"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No data found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/top10/tv-shows/{country}/{provider}:
    get:
      operationId: getTop10TVShowsByProvider
      summary: Get Top 10 TV shows for a provider
      parameters:
        - $ref: "#/components/parameters/CountryParam"
        - $ref: "#/components/parameters/ProviderParam"
        - $ref: "#/components/parameters/DateParam"
      responses:
        "200":
          description: Top 10 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Top10Response"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No data found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/top10/{country}:
    get:
      operationId: getCountryTop10
      summary: Every provider's movie and TV Top 10 of a country in one response
      parameters:
        - $ref: "#/components/parameters/CountryParam"
        - $ref: "#/components/parameters/DateParam"
      responses:
        "200":
          description: Charts (each provider's latest chart unless date is given)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Top10Combined"
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/top10/global/{provider}:
    get:
      operationId: getGlobalTop10
      summary: A provider's titles ranked across every country by chart points
      description: >
        Points are FlixPatrol-style (rank 1 = 10 points … rank 10 = 1) summed
        over every stored country chart of the day. Titles mapped to the same
        TMDB title are one entry.
      parameters:
        - $ref: "#/components/parameters/ProviderParam"
        - $ref: "#/components/parameters/DateParam"
        - $ref: "#/components/parameters/CategoryQuery"
        - $ref: "#/components/parameters/KindQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Leaderboard of one day
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/top10/movies/{country}/{provider}/history:
    get:
      operationId: getMovieChartHistory
      summary: One movie chart over a date range
      parameters:
        - $ref: "#/components/parameters/CountryParam"
        - $ref: "#/components/parameters/ProviderParam"
        - $ref: "#/components/parameters/FromParam"
        - $ref: "#/components/parameters/ToParam"
      description: Defaults to the last 7 days; at most 92 days. Only stored dates are listed.
      responses:
        "200":
          description: Chart per stored date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartHistoryResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/top10/tv-shows/{country}/{provider}/history:
    get:
      operationId: getTVChartHistory
      summary: One TV chart over a date range
      parameters:
        - $ref: "#/components/parameters/CountryParam"
        - $ref: "#/components/parameters/ProviderParam"
        - $ref: "#/components/parameters/FromParam"
        - $ref: "#/components/parameters/ToParam"
      description: Defaults to the last 7 days; at most 92 days. Only stored dates are listed.
      responses:
        "200":
          description: Chart per stored date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartHistoryResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/charts:
    get:
      operationId: listCharts
      summary: Every stored chart with its date range and freshness
      parameters:
        - name: country
          in: query
          schema: { type: string, description: ISO country code }
        - name: provider
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Chart catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartCatalogResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/charts/{country}/{provider}/dates:
    get:
      operationId: listChartDates
      summary: Dates stored for a chart, newest first
      parameters:
        - $ref: "#/components/parameters/CountryParam"
        - $ref: "#/components/parameters/ProviderParam"
        - $ref: "#/components/parameters/CategoryQuery"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 1000, default: 366 }
      responses:
        "200":
          description: Stored dates
          content:
            application/json:
              schema:
                type: object
                properties:
                  country: { type: string }
                  provider: { type: string }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        date: { type: string, format: date }
                        category: { type: string, enum: [movies, tv_shows] }
                        entries: { type: integer }
                        scraped_at: { type: string, format: date-time }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/changes:
    get:
      operationId: listChartChanges
      summary: Charts stored or re-scraped after a time (poll this instead of every chart)
      description: >
        Pass the response's `next` as `since` on the next call. Default since:
        24 hours ago.
      parameters:
        - name: since
          in: query
          schema: { type: string, format: date-time, example: "2026-10-09T12:00:00Z" }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
      responses:
        "200":
          description: Changed charts, oldest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartChangesResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/movers:
    get:
      operationId: getMovers
      summary: Climbers, fallers, debuts, re-entries and exits of a chart
      description: >
        Compares the chart on `date` (default: the latest stored date) with
        the stored date before it. Debuts are titles not on the chart in the
        previous 90 days.
      parameters:
        - name: country
          in: query
          required: true
          schema: { type: string }
        - name: provider
          in: query
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/CategoryQuery"
        - $ref: "#/components/parameters/DateParam"
      responses:
        "200":
          description: Movers per category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoversResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/leaderboards:
    get:
      operationId: getLeaderboard
      summary: Titles ranked by chart points over a period
      description: >
        Points per chart per day: rank 1 = 10 … rank 10 = 1. Omit country /
        provider to sum over all. period=week (default) is the 7 days ending
        on `date` (default: the current chart date); custom uses from/to (at
        most 366 days).
      parameters:
        - name: period
          in: query
          schema: { type: string, enum: [day, week, month, custom], default: week }
        - $ref: "#/components/parameters/DateParam"
        - $ref: "#/components/parameters/FromParam"
        - $ref: "#/components/parameters/ToParam"
        - $ref: "#/components/parameters/HistoryCountryParam"
        - name: provider
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/CategoryQuery"
        - $ref: "#/components/parameters/KindQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
    BearerKey:
      type: http
      scheme: bearer
      description: 'The same API key, as "Authorization: Bearer <key>"'
...
  parameters:
...
    TmdbKindParam:
      name: kind
      in: path
      required: true
      schema:
        type: string
        enum: [movie, tv]
        description: TMDB namespace of the ID
    HistoryCountryParam:
      name: country
      in: query
      required: false
      schema:
        type: string
        description: ISO country code (e.g. US); omit for all countries
    FromParam:
      name: from
      in: query
      required: false
      schema:
        type: string
        format: date
        description: First chart date to include (YYYY-MM-DD)
    ToParam:
      name: to
      in: query
      required: false
      schema:
        type: string
        format: date
        description: Last chart date to include (YYYY-MM-DD)
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
      required: [error]
...
    TitlesPageResponse:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TitleResponse"
        total:
          type: integer
          format: int64
        limit:
          type: integer
          format: int64
        offset:
          type: integer
          format: int64
      required: [items, total, limit, offset]
...
    ServicesResponse:
      type: object
      additionalProperties: false
      properties:
        country:
          type: string
        services:
          type: array
          items:
            $ref: "#/components/schemas/ReleaseService"
      required: [country, services]
...